requirements.txt
flask>=3.0.0
openpyxl>=3.1.0
python-pptx>=1.0.0
pypdf>=5.0.0
requests>=2.31.0
